Added Authentication

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -47,11 +47,20 @@ public class JournalEntryService {
     public Optional<JournalEntry> findById(ObjectId myId) {
         return journalEntryRepository.findById(myId);
     }
-
-    public void deleteById(ObjectId myId, String userName) {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(myId));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(myId);
+@Transactional
+    public boolean deleteById(ObjectId myId, String userName) {
+    boolean removed = false;
+       try{
+           User user = userService.findByUserName(userName);
+           removed = user.getJournalEntries().removeIf(x -> x.getId().equals(myId));
+           if(removed) {
+               userService.saveEntry(user);
+               journalEntryRepository.deleteById(myId);
+           }
+       }catch(Exception e) {
+           System.out.println(e.getMessage());
+           throw new RuntimeException("An error occurred while deleting the entry", e);
+       }
+       return removed;
     }
 }
